api-requirements.txt
# API-specific requirements
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.9.0
psycopg2-binary>=2.9.3
boto3>=1.24.0
ipaddress>=1.0.0
python-multipart>=0.0.5
httpx>=0.24.0
pytest>=6.2.0
pandas>=1.3.0